flask_app/requirements.txt
Flask==2.3.2
pyserial==3.5
Flask-Cors==3.1.3
